requirements.txt
alembic==1.4.2
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
ecdsa==0.13
email-validator==1.1.1
Flask==1.1.2
Flask-Caching==1.7.1
Flask-Login==0.5.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.0
MarkupSafe==1.1.1
olefile==0.46
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.9.7
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
stripe==2.36.2
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
gunicorn==20.0.4
